refactor: main components를 세부 폴더로 분리
