+++ src/app/[locale]/home/components/header/SearchPlace.tsx
