+++ src/app/[locale]/home/components/place/PlaceList.tsx
