+++ src/mocks/handlers/place.ts
@@ -1,9 +1,9 @@
 import { HttpResponse, http } from "msw";
 
-import { IResponsePlace } from "@/lib/types/place";
+import { ResponsePlace } from "@/lib/types/place";
 import { placeResponse } from "../response/place";
 
-export const getPlaces = http.get<never, never, IResponsePlace>(
+export const getPlaces = http.get<never, never, ResponsePlace>(
   `${process.env.API_HOST}/place`,
   () => {
     return HttpResponse.json(placeResponse);
